move toolbar alarms to its folder
